Split Python impl into multiple files
