+++ docs/api.md
@@ -1,3 +1,3 @@
 # API Reference
 
-::: py_mini_racer.py_mini_racer
+::: py_mini_racer
