+++ src/py_mini_racer/_context_base.py
@@ -1,0 +1,120 @@
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import (
+    TYPE_CHECKING,
+    Callable,
+)
+
+from py_mini_racer._types import JSUndefined
+
+if TYPE_CHECKING:
+    from py_mini_racer._numeric import Numeric
+    from py_mini_racer._objects import (
+        JSArray,
+        JSFunction,
+        JSPromise,
+    )
+    from py_mini_racer._types import (
+        JSEvalException,
+        JSObject,
+        JSUndefinedType,
+        PythonJSConvertedTypes,
+    )
+
+
+class ValueHandleBase(ABC):
+    @property
+    @abstractmethod
+    def raw(self) -> object:
+        pass
+
+    @abstractmethod
+    def to_python(self) -> PythonJSConvertedTypes | JSEvalException:
+        pass
+
+    @abstractmethod
+    def to_python_or_raise(self) -> PythonJSConvertedTypes:
+        pass
+
+
+class ContextBase(ABC):
+    """A Context provides Pythonic wrappers around the MiniRacer C API.
+
+    This is intended for internal usage by py_mini_racer. MiniRacerContext provides
+    a further wrapper around this interface.
+    """
+
+    @abstractmethod
+    def get_identity_hash(self, obj: JSObject) -> int:
+        pass
+
+    @abstractmethod
+    def get_own_property_names(
+        self, obj: JSObject
+    ) -> tuple[PythonJSConvertedTypes, ...]:
+        pass
+
+    @abstractmethod
+    def get_object_item(
+        self, obj: JSObject, key: PythonJSConvertedTypes
+    ) -> PythonJSConvertedTypes:
+        pass
+
+    @abstractmethod
+    def set_object_item(
+        self, obj: JSObject, key: PythonJSConvertedTypes, val: PythonJSConvertedTypes
+    ) -> None:
+        pass
+
+    @abstractmethod
+    def del_object_item(self, obj: JSObject, key: PythonJSConvertedTypes) -> None:
+        pass
+
+    @abstractmethod
+    def del_from_array(self, arr: JSArray, index: int) -> None:
+        pass
+
+    @abstractmethod
+    def array_insert(
+        self, arr: JSArray, index: int, new_val: PythonJSConvertedTypes
+    ) -> None:
+        pass
+
+    @abstractmethod
+    def call_function(
+        self,
+        func: JSFunction,
+        *args: PythonJSConvertedTypes,
+        this: JSObject | JSUndefinedType = JSUndefined,
+        timeout_sec: Numeric | None = None,
+    ) -> PythonJSConvertedTypes:
+        pass
+
+    @abstractmethod
+    def make_js_callback(
+        self, func: Callable[[PythonJSConvertedTypes | JSEvalException], None]
+    ) -> tuple[Callable[[], None], JSFunction]:
+        pass
+
+    @abstractmethod
+    def promise_then(
+        self, promise: JSPromise, on_resolved: JSFunction, on_rejected: JSFunction
+    ) -> None:
+        pass
+
+    @abstractmethod
+    def create_intish_val(self, val: int, typ: int) -> ValueHandleBase:
+        pass
+
+    @abstractmethod
+    def create_doublish_val(self, val: float, typ: int) -> ValueHandleBase:
+        pass
+
+    @abstractmethod
+    def create_string_val(self, val: str, typ: int) -> ValueHandleBase:
+        pass
+
+    @abstractmethod
+    def free(self, val_handle: ValueHandleBase) -> None:
+        pass
