+++ src/py_mini_racer/_numeric.py
@@ -1,0 +1,5 @@
+from typing import (
+    Union,
+)
+
+Numeric = Union[int, float]
